+++ v3/crates/metadata-resolve/src/stages/type_permissions/error.rs
@@ -21,12 +21,6 @@ pub enum TypeOutputPermissionError {
         field_name: FieldName,
         type_name: Qualified<CustomTypeName>,
     },
-    #[error(
-        "Output TypePermissions for object type {object_type_name} use authorization rules but they are not enabled"
-    )]
-    AuthorizationRulesNotEnabled {
-        object_type_name: Qualified<CustomTypeName>,
-    },
 }
 
 impl ContextualError for TypeOutputPermissionError {
@@ -61,12 +55,6 @@ pub enum TypeInputPermissionError {
         field_name: FieldName,
         type_name: CustomTypeName,
         type_error: typecheck::TypecheckError,
-    },
-    #[error(
-        "Input TypePermissions for object type {object_type_name} use authorization rules but they are not enabled"
-    )]
-    AuthorizationRulesNotEnabled {
-        object_type_name: Qualified<CustomTypeName>,
     },
 }