+++ v3/crates/metadata-resolve/tests/metadata_golden_tests.rs
@@ -91,7 +91,6 @@ fn read_test_configuration(
 ) -> Result<configuration::Configuration, Box<dyn std::error::Error>> {
     let unstable_features = configuration::UnstableFeatures {
         enable_aggregation_predicates: true,
-        enable_authorization_rules: true,
     };
 
     let configuration_path = directory.join("configuration.json");