+++ v3/crates/open-dds/src/authorization.rs
@@ -57,14 +57,14 @@ pub struct Comparison {
 #[opendd(externally_tagged, json_schema(title = "TypeAuthorizationRule"))]
 /// A rule that determines which fields of a type are available to a user
 pub enum TypeAuthorizationRule {
-    // if a condition is provided, it must evaluate to `true` for these fields
+    // If a condition is provided, it must evaluate to `true` for these fields
     // to be made available to the user
     AllowFields(Fields),
-    // if a condition is provided, it must evaluate to `true` for these fields
+    // If a condition is provided, it must evaluate to `true` for these fields
     // to be denied to the user. A denied field takes precedence over an allowed field.
     DenyFields(Fields),
-    // if a condition is provided, it must evaluate to `true` for this field
-    // to be preset for the user.
+    // If a condition is provided, it must evaluate to `true` for this field
+    // to be preset for the user. Where multiple entries for the same field match the last one wins.
     FieldPreset(InputTypeFieldPreset),
 }
 
@@ -72,6 +72,7 @@ pub enum TypeAuthorizationRule {
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
 #[opendd(json_schema(title = "Fields"))]
+/// A list of fields to allow/deny when the condition evaluates to `true`
 pub struct Fields {
     pub fields: Vec<FieldName>,
     pub condition: Option<Condition>,
@@ -81,6 +82,7 @@ pub struct Fields {
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
 #[opendd(json_schema(title = "InputTypeFieldPreset"))]
+/// A preset value for a field that applies when the condition evaluates to `true`
 pub struct InputTypeFieldPreset {
     pub condition: Option<Condition>,
     pub field_name: FieldName,
@@ -93,12 +95,14 @@ pub struct InputTypeFieldPreset {
 #[opendd(externally_tagged, json_schema(title = "CommandAuthorizationRule"))]
 /// A rule that determines which commands and argument presets are available to a user
 pub enum CommandAuthorizationRule {
-    // if a condition is provided, it must evaluate to 'true' for
+    // If a condition is provided, it must evaluate to 'true' for
     // this Command to be available to the user
     Allow(Allow),
-    // if the provided condition evaluates to 'true' this Command will not be available to the user
+    // If the provided condition evaluates to 'true' this Command will not be available to the user
     Deny(Deny),
-    // if a condition is provided, it must evaluate to `true` for this argument to be preset
+    // If a condition is provided, it must evaluate to `true` for this argument to be preset.
+    // Multiple boolean expression arguments that match will be combined with `AND`, in the case of other multiple
+    // arguments the last one wins.
     PresetArgument(PresetArgument),
 }
 
@@ -106,6 +110,7 @@ pub enum CommandAuthorizationRule {
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
 #[opendd(json_schema(title = "Allow"))]
+/// Allow access if the condition evaluates to `true`
 pub struct Allow {
     pub condition: Option<Condition>,
 }
@@ -113,6 +118,7 @@ pub struct Allow {
 #[derive(Serialize, Clone, Debug, PartialEq, Eq, opendds_derive::OpenDd)]
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
+/// Deny access if the condition evaluates to `true`
 #[opendd(json_schema(title = "Deny"))]
 pub struct Deny {
     pub condition: Condition,
@@ -122,6 +128,7 @@ pub struct Deny {
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
 #[opendd(json_schema(title = "PresetArgument"))]
+/// A preset value for an argument that applies when the condition evaluates to `true`
 pub struct PresetArgument {
     pub argument_name: Spanned<ArgumentName>,
     pub condition: Option<Condition>,
@@ -134,20 +141,22 @@ pub struct PresetArgument {
 #[opendd(externally_tagged, json_schema(title = "ModelAuthorizationRule"))]
 /// A rule that determines which model and argument presets are available to a user
 pub enum ModelAuthorizationRule {
-    // if a condition is provided, it must evaluate to 'true' for
+    // If a condition is provided, it must evaluate to 'true' for
     // this Model to be available to the user
     Allow(Allow),
-    // if the provided condition evaluates to 'true' this Model will not be available to the user
+    // If the provided condition evaluates to 'true' this Model will not be available to the user
     Deny(Deny),
-    // if a condition is provided, it must evaluate to `true` for this argument to be preset
+    // If a condition is provided, it must evaluate to `true` for this argument to be preset.
+    // Multiple boolean expression arguments that match will be combined with `AND`, in the case of other multiple
+    // arguments the last one wins.
     PresetArgument(PresetArgument),
-    // if a condition is provided, it must evaluate to `true` for this filter to be included in
-    // requests. Multiple filters that match will be combined with `AND`
+    // If a condition is provided, it must evaluate to `true` for this filter to be included in
+    // requests. Multiple filters that match will be combined with `AND`.
     Filter(Filter),
-    // if a condition is provided, it must evaluate to `true` for these relational operations to be
+    // If a condition is provided, it must evaluate to `true` for these relational operations to be
     // available to the user
     AllowRelationalOperations(AllowRelationalOperations),
-    // if the provided condition evaluates to `true` these relational operations will not be available
+    // If the provided condition evaluates to `true` these relational operations will not be available
     // to the user
     DenyRelationalOperations(DenyRelationalOperations),
 }
@@ -156,6 +165,7 @@ pub enum ModelAuthorizationRule {
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
 #[opendd(json_schema(title = "Filter"))]
+/// A filter that applies when the condition evaluates to `true`
 pub struct Filter {
     pub condition: Option<Condition>,
     pub predicate: ModelPredicate,
@@ -166,6 +176,7 @@ pub struct Filter {
 )]
 #[schemars(title = "RelationalOperation")]
 #[serde(rename_all = "camelCase")]
+/// An operation on a model
 pub enum RelationalOperation {
     Insert,
     Update,
@@ -176,6 +187,7 @@ pub enum RelationalOperation {
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
 #[opendd(json_schema(title = "AllowRelationalOperations"))]
+/// Allow these relational operations if the condition evaluates to `true`
 pub struct AllowRelationalOperations {
     pub condition: Option<Condition>,
     pub operations: Vec<RelationalOperation>,
@@ -185,6 +197,7 @@ pub struct AllowRelationalOperations {
 #[serde(rename_all = "camelCase")]
 #[serde(deny_unknown_fields)]
 #[opendd(json_schema(title = "DenyRelationalOperations"))]
+/// Deny these relational operations if the condition evaluates to `true`
 pub struct DenyRelationalOperations {
     pub condition: Condition,
     pub operations: Vec<RelationalOperation>,